feat(tree_search_strategy): add optional multiprocessing pool
